requirements.txt
asgiref==3.7.2
Django==4.2.1
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.4
tzdata==2023.3
wheel==0.38.4
Pillow
djangorestframework~=3.14.0
drf-yasg
djangorestframework-simplejwt
pytz~=2023.3